python/gui/src/Saphir/requirements.txt
PySide6
pyserial
watchdog
